requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
pydantic
fastapi
uvicorn[standard]
supabase
httpx
